ecn_conflict_demo/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq>=0.13.0
python-dotenv==1.0.1
networkx==3.3
pydantic==2.7.1
